+++ test/services/school/index.test.js
@@ -174,7 +174,7 @@ describe('school service', () => {
 		});
 	});
 
-	describe('patch schools', () => {
+	describe.only('patch schools', () => {
 		it('administrator can patch his own school', async () => {
 			const school = await testObjects.createTestSchool({});
 			const admin = await testObjects.createTestUser({
@@ -341,6 +341,54 @@ describe('school service', () => {
 				expect(result.isTeamCreationByStudentsEnabled).to.be.false;
 			}
 		);
+		it('should fail to update officialSchoolNumber with wrong format', async () => {
+			const school = await testObjects.createTestSchool({});
+			const admin = await testObjects.createTestUser({
+				schoolId: school._id,
+				roles: ['administrator'],
+			});
+			const params = await testObjects.generateRequestParamsFromUser(admin);
+
+			try {
+				await app.service('/schools').patch(school._id, { officialSchoolNumber: 'foo' }, params);
+			} catch (err) {
+				expect(err).to.not.equal(undefined);
+				expect(err.message).to.include('School ID is incorrect');
+				expect(err.name).to.be.equal('Error');
+			}
+		});
+		it('should succeed to update officialSchoolNumber with correct format', async () => {
+			const school = await testObjects.createTestSchool({});
+			const admin = await testObjects.createTestUser({
+				schoolId: school._id,
+				roles: ['administrator'],
+			});
+			const params = await testObjects.generateRequestParamsFromUser(admin);
+
+			const mockOfficialSchoolNumber = 'BA-13372';
+			let result;
+
+			try {
+				result = await app
+					.service('/schools')
+					.patch(school._id, { officialSchoolNumber: mockOfficialSchoolNumber }, params);
+			} catch (err) {
+				console.error(err);
+			}
+			expect(result.officialSchoolNumber).to.be.equal(mockOfficialSchoolNumber);
+
+			const mockOfficialSchoolNumber2 = '13372';
+			let result2;
+
+			try {
+				result2 = await app
+					.service('/schools')
+					.patch(school._id, { officialSchoolNumber: mockOfficialSchoolNumber2 }, params);
+			} catch (err) {
+				console.error(err);
+			}
+			expect(result2.officialSchoolNumber).to.be.equal(mockOfficialSchoolNumber2);
+		});
 	});
 });